Project 2: Custom Exception Handling
This java project asks user for his/her NSU ID and checks if it's valid or not by counting its digits. If its not a valid ID then it throws a custom exception.

This project contains a custom coded Exception Class which occurs when that custom exception is thrown and create an exception object. This exception will ask for the user's valid id until its correct.

After giving the correct id, the program will then provide information about the admission date and semester of admision of that user.